requirements.txt
flask
boto3
langchain
pypdf
faiss-cpu
python-dotenv
requests
